01-04

+++ Web/Node.js/SocketServer/src/client/script/login.js
@@ -1,26 +1,40 @@
 'use strict';
 
-const socketPort = 3000;
-
 const $ = (e) => document.querySelector(e);
 const $$ = (e) => document.querySelectorAll(e);
 
 const username = $('#username');
 const form = $('form');
+const url = 'http://localhost:3000';
+
+const config = (data) => {
+  return {
+    body: JSON.stringify(data),
+    headers: {'content-type': 'application/json'},
+    method: 'POST',
+  };
+};
 
 async function submit() {
   const name = username.value;
 
-  await (
-    await fetch(`http://localhost:${socketPort}/login`, {
-      body: JSON.stringify({name}),
-      headers: {'content-type': 'application/json'},
-      method: 'POST',
-    })
-  )
+  await (await fetch(`${url}/login`, config({name, register: false})))
     .json()
-    .then((res) => {
-      localStorage.setItem('token', res);
+    .then(async (res) => {
+      let thisToken = res.token;
+      if (res.newUser) {
+        if (confirm('username not found, sure continue to register?')) {
+          await (await fetch(`${url}/login`, config({name, register: true})))
+            .json()
+            .then((res) => {
+              thisToken = res.token;
+            });
+        } else {
+          username.value = '';
+          return;
+        }
+      }
+      localStorage.setItem('token', thisToken);
       localStorage.setItem('curUser', name);
 
       window.location = '../client/';
@@ -29,5 +43,5 @@ async function submit() {
 
 form.onsubmit = $('#submit').onclick = () => submit();
 username.onkeydown = (e) => {
-  if (e.key === 'Enter') submit();
+  if (e.key === 'Enter') submit().then();
 };
